fix(face-recognizer): address PR review findings
- deploy.sh: fix binary path from face-recognizer/src/cpp/ to the correct
  CMake output directory face-recognizer_cpp/; fix ArcFace model filename
  from w600k_mbf to w600k_r50; rewrite config.yaml paths after deploy so
  model and gallery paths resolve from the deploy directory rather than
  apps root; strip sink:display so --test works headless out of the box.
- main.cpp: guard output_sink YAML fallback with output_sink_explicit flag
  so --test and --no-display are not overridden by config.yaml sink:display;
  add centroid-drift check to cached match reuse so detection reordering
  (SCRFD sorts by score) triggers fresh ArcFace recognition on multi-face scenes.
- scrfd_decode.cpp: cast fh to size_t before anchor grid multiplication to
  prevent signed int overflow before implicit conversion to size_type (CodeQL).

+++ examples/face-recognition/face-recognizer/src/cpp/scrfd_decode.cpp
@@ -579,7 +579,7 @@ static std::vector<Anchor> make_anchors(int scale_idx, int infer_h, int infer_w,
     const int fh = static_cast<int>(std::ceil(static_cast<float>(infer_h) / stride));
     const int fw = static_cast<int>(std::ceil(static_cast<float>(infer_w) / stride));
     std::vector<Anchor> anchors;
-    anchors.reserve(fh * fw * num_anchors);
+    anchors.reserve(static_cast<size_t>(fh) * fw * num_anchors);
     for (int i = 0; i < fh; ++i)
         for (int j = 0; j < fw; ++j)
             for (int a = 0; a < num_anchors; ++a)
